+++ library/android-mvc/src/main/java/com/shipdream/lib/android/mvc/view/MvcActivity.java
@@ -350,6 +350,11 @@ public abstract class MvcActivity extends AppCompatActivity {
             activity.delegateFragment = this;
         }
 
+        void onPreViewReady(final View view, final Bundle savedInstanceState) {
+            if (savedInstanceState != null) {
+                notifyAllSubMvcFragmentsTheirStateIsManagedByMe(this, true);
+            }
+        }
 
         private boolean firstTimeRun = false;
 
@@ -359,10 +364,6 @@ public abstract class MvcActivity extends AppCompatActivity {
             canCommitFragmentTransaction = true;
             if (reason.isFirstTime()) {
                 firstTimeRun = true;
-            }
-
-            if (savedInstanceState != null) {
-                notifyAllSubMvcFragmentsTheirStateIsManagedByMe(this, true);
             }
         }
 
@@ -439,6 +440,8 @@ public abstract class MvcActivity extends AppCompatActivity {
             DefaultStateKeeperHolder.saveStateOfAllControllers(mvcOutState);
             outState.putBundle(MVC_STATE_BUNDLE_KEY, mvcOutState);
             logger.trace("Save state of all active controllers, {}ms used.", System.currentTimeMillis() - ts);
+
+            notifyAllSubMvcFragmentsTheirStateIsManagedByMe(this, true);
         }
 
         /**
@@ -502,7 +505,6 @@ public abstract class MvcActivity extends AppCompatActivity {
                 final MvcFragment fragment;
                 try {
                     fragment = (MvcFragment) new ReflectUtils.newObjectByType(fragmentClass).newInstance();
-                    fragment.injectDependencies();
                 } catch (Exception e) {
                     throw new RuntimeException("Failed to instantiate fragment: " + fragmentClass.getName(), e);
                 }
@@ -598,29 +600,12 @@ public abstract class MvcActivity extends AppCompatActivity {
                         }
                     }
 
-                    MvcFragment lastFrag = null;
-                    if (event.getLastValue() != null && event.getLastValue().getLocationId() != null) {
-                        lastFrag = (MvcFragment) fm.findFragmentByTag(
-                                getFragmentTag(event.getLastValue().getLocationId()));
-                    }
-                    final MvcFragment finalLastFrag = lastFrag;
-
-                    if (finalLastFrag != null) {
-                        finalLastFrag.selfRelease = false;
-                    }
-
                     currentFrag.registerOnViewReadyListener(new Runnable() {
                         @Override
                         public void run() {
                             if (event.getNavigator() != null) {
                                 __MvcControllerHelper.destroyNavigator(event.getNavigator());
                             }
-
-                            if (finalLastFrag != null) {
-                                finalLastFrag.releaseDependencies();
-                                finalLastFrag.selfRelease = true;
-                            }
-
                             currentFrag.unregisterOnViewReadyListener(this);
                         }
                     });